+++ abilian/web/admin/panels/users/views.py
@@ -81,7 +81,8 @@ class JsonUsersList(base.JSONView):
       mugshot = user_photo_url(user, size=MUGSHOT_SIZE)
       name = escape(getattr(user, "name") or "")
       email = escape(getattr(user, "email") or "")
-      roles = security.get_roles(user, no_group_roles=True)
+      roles = [r for r in security.get_roles(user, no_group_roles=True)
+               if r.assignable]
       columns = []
       columns.append(
         u'<a href="{url}"><img src="{src}" width="{size}" height="{size}">'
@@ -142,7 +143,7 @@ class UserEdit(UserBase, views.ObjectEdit):
     kw = super(UserEdit, self).get_form_kwargs()
     security = current_app.services['security']
     roles = security.get_roles(self.obj, no_group_roles=True)
-    kw['roles'] = [r.name for r in roles]
+    kw['roles'] = [r.name for r in roles if r.assignable]
     return kw
 
   def validate(self):
@@ -167,7 +168,8 @@ class UserEdit(UserBase, views.ObjectEdit):
 
   def after_populate_obj(self):
     security = current_app.services['security']
-    current_roles = set(security.get_roles(self.obj, no_group_roles=True))
+    current_roles = security.get_roles(self.obj, no_group_roles=True)
+    current_roles = set(r for r in current_roles if r.assignable)
     new_roles = {Role(r) for r in self.form.roles.data}
 
     for r in (current_roles - new_roles):